Moved ILoggerAdapter and LoggerAdapter to their specific assemblies and ran a code refactor.
