+++ StockTracker.Adapter.Interface/StockTracker.Adapter.Interface.csproj
@@ -4,8 +4,4 @@
     <TargetFramework>netcoreapp2.1</TargetFramework>
   </PropertyGroup>
 
-  <ItemGroup>
-    <Folder Include="Logger\" />
-  </ItemGroup>
-
 </Project>
